+++ src/main/java/org/noostak/group/application/GroupCreateService.java
@@ -1,7 +1,7 @@
 package org.noostak.group.application;
 
 import org.noostak.group.dto.request.GroupCreateRequest;
-import org.noostak.group.dto.response.GroupCreateInternalResponse;
+import org.noostak.group.dto.response.create.GroupCreateInternalResponse;
 
 import java.io.IOException;
