+++ src/main/java/org/noostak/group/application/GroupRetrieveService.java
@@ -1,6 +1,6 @@
 package org.noostak.group.application;
 
-import org.noostak.group.dto.response.GroupsRetrieveResponse;
+import org.noostak.group.dto.response.retrieve.GroupsRetrieveResponse;
 
 public interface GroupRetrieveService {
     GroupsRetrieveResponse findGroups(Long memberId);
